Move the rest of classes from the lib directory to the Service dir
