+++ Command/GenerateFixturesCommand.php
@@ -2,8 +2,13 @@
 
 namespace FixtureBundle\Command;
 
+use FixtureBundle\Repository\FolderRepository;
 use FixtureBundle\Service\FixtureLoader;
+use FixtureBundle\Service\Generator;
 use Pimcore\Console\AbstractCommand;
+use Pimcore\Model\Element\AbstractElement;
+use Pimcore\Model\Object\AbstractObject;
+use Pimcore\Model\Object\Folder;
 use Symfony\Bundle\FrameworkBundle\Command\ContainerAwareCommand;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
@@ -11,6 +16,7 @@ use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Question\ChoiceQuestion;
 use Symfony\Component\Console\Question\ConfirmationQuestion;
+use Symfony\Component\Console\Question\Question;
 use Symfony\Component\Validator\Constraints\Choice;
 
 class GenerateFixturesCommand extends AbstractCommand
@@ -41,7 +47,6 @@ class GenerateFixturesCommand extends AbstractCommand
 
         $rootFolder = $helper->ask($input, $output, $folderRootQuestion);
 
-
         $filenameQuestion = new Question('<info>Choose filename: </info>', 'test');
         $levelsQuestion = new Question('<info>Choose max levels deep (100): </info>', 100);
 
@@ -64,6 +69,7 @@ class GenerateFixturesCommand extends AbstractCommand
         if (!$helper->ask($input, $output, $confirmationQuestion)) {
             return;
         }
+
         foreach (glob(FixtureLoader::FIXTURE_FOLDER . '_generated' . DIRECTORY_SEPARATOR . '*.yml') as $file){
             unlink($file);
         }
